# api/api/ai/model/v1/model_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: ai/model/v1/model.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationModelGetDefaultModel = "/ai.model.v1.Model/GetDefaultModel"
const OperationModelGetModel = "/ai.model.v1.Model/GetModel"
const OperationModelListModel = "/ai.model.v1.Model/ListModel"

type ModelHTTPServer interface {
	GetDefaultModel(context.Context, *DefaultModelRequest) (*GetModelResponse, error)
	GetModel(context.Context, *SimpleRequest) (*GetModelResponse, error)
	ListModel(context.Context, *ListModelRequest) (*ListModelResponse, error)
}

func RegisterModelHTTPServer(s *http.Server, srv ModelHTTPServer) {
	r := s.Route("/")
	r.POST("/ai/model/list/me", _Model_ListModel1_HTTP_Handler(srv))
	r.GET("/ai/model/{id}", _Model_GetModel1_HTTP_Handler(srv))
	r.POST("/ai/model/default", _Model_GetDefaultModel0_HTTP_Handler(srv))
}

func _Model_ListModel1_HTTP_Handler(srv ModelHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListModelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelListModel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListModel(ctx, req.(*ListModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListModelResponse)
		return ctx.Result(200, reply)
	}
}

func _Model_GetModel1_HTTP_Handler(srv ModelHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SimpleRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelGetModel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetModel(ctx, req.(*SimpleRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetModelResponse)
		return ctx.Result(200, reply)
	}
}

func _Model_GetDefaultModel0_HTTP_Handler(srv ModelHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DefaultModelRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationModelGetDefaultModel)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetDefaultModel(ctx, req.(*DefaultModelRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetModelResponse)
		return ctx.Result(200, reply)
	}
}

type ModelHTTPClient interface {
	GetDefaultModel(ctx context.Context, req *DefaultModelRequest, opts ...http.CallOption) (rsp *GetModelResponse, err error)
	GetModel(ctx context.Context, req *SimpleRequest, opts ...http.CallOption) (rsp *GetModelResponse, err error)
	ListModel(ctx context.Context, req *ListModelRequest, opts ...http.CallOption) (rsp *ListModelResponse, err error)
}

type ModelHTTPClientImpl struct {
	cc *http.Client
}

func NewModelHTTPClient(client *http.Client) ModelHTTPClient {
	return &ModelHTTPClientImpl{client}
}

func (c *ModelHTTPClientImpl) GetDefaultModel(ctx context.Context, in *DefaultModelRequest, opts ...http.CallOption) (*GetModelResponse, error) {
	var out GetModelResponse
	pattern := "/ai/model/default"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationModelGetDefaultModel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ModelHTTPClientImpl) GetModel(ctx context.Context, in *SimpleRequest, opts ...http.CallOption) (*GetModelResponse, error) {
	var out GetModelResponse
	pattern := "/ai/model/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationModelGetModel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ModelHTTPClientImpl) ListModel(ctx context.Context, in *ListModelRequest, opts ...http.CallOption) (*ListModelResponse, error) {
	var out ListModelResponse
	pattern := "/ai/model/list/me"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationModelListModel))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}