# api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /ai/admin/model:
        post:
            tags:
                - Admin
            description: Model
            operationId: Admin_CreateModel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.common.v1.AiModel'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.common.v1.AiModel'
...
    /ai/admin/model/{id}:
        get:
            tags:
                - Admin
            operationId: Admin_GetModel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.common.v1.AiModel'
        put:
            tags:
                - Admin
            operationId: Admin_UpdateModel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.common.v1.AiModel'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.common.v1.AiModel'
        delete:
            tags:
                - Admin
            operationId: Admin_DeleteModel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
...
    /ai/model/default:
        post:
            tags:
                - Model
            operationId: Model_GetDefaultModel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.model.v1.DefaultModelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.model.v1.GetModelResponse'
    /ai/model/list/me:
        post:
            tags:
                - Model
            operationId: Model_ListModel
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ai.model.v1.ListModelRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.model.v1.ListModelResponse'
    /ai/model/{id}:
        get:
            tags:
                - Model
            operationId: Model_GetModel
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ai.model.v1.GetModelResponse'
...
components:
    schemas:
...
        ai.common.v1.AiApiKey:
            type: object
            properties:
                id:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                deletedAt:
                    type: string
                    format: date-time
                name:
                    type: string
                apiKey:
                    type: string
                platform:
                    type: string
                url:
                    type: string
                status:
                    type: string
                aiModel:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.common.v1.AiModel'
...
        ai.common.v1.AiModel:
            type: object
            properties:
                id:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
                deletedAt:
                    type: string
                    format: date-time
                name:
                    type: string
                type:
                    type: string
                platform:
                    type: string
                sort:
                    type: string
                status:
                    type: string
                temperature:
                    type: number
                    format: double
                maxTokens:
                    type: string
                maxContexts:
                    type: string
                keyId:
                    type: string
                model:
                    type: string
                aiApiKey:
                    $ref: '#/components/schemas/ai.common.v1.AiApiKey'
...
        ai.model.v1.DefaultModelRequest:
            type: object
            properties:
                type:
                    type: string
                platform:
                    type: string
        ai.model.v1.GetModelResponse:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                type:
                    type: string
                platform:
                    type: string
                temperature:
                    type: number
                    format: double
                maxTokens:
                    type: integer
                    format: int32
                maxContexts:
                    type: integer
                    format: int32
                model:
                    type: string
        ai.model.v1.ListModelRequest:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationArgs'
                name:
                    type: string
                type:
                    type: string
                platform:
                    type: string
        ai.model.v1.ListModelResponse:
            type: object
            properties:
                models:
                    type: array
                    items:
                        $ref: '#/components/schemas/ai.model.v1.GetModelResponse'
                pagination:
                    $ref: '#/components/schemas/common.v1.PaginationResults'
...
        common.v1.PaginationArgs:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                orderBy:
                    type: string
                orderDirection:
                    type: string
        common.v1.PaginationResults:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                totalItems:
                    type: integer
                    format: int32
...